water_tracker/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Werkzeug==2.3.7
matplotlib==3.8.0
email-validator==2.1.0
opencv-python==4.8.0.74
pytesseract==0.3.10
SpeechRecognition==3.10.0
pydub==0.25.1
numpy==1.24.3
Pillow==10.0.0
customtkinter==5.2.0
